Input content for landing page
However, whether I've added the content in the right way remains
to be seen. I suspect I will need to include more <div> containers
when it comes to the styling.

Something I really need to look into is when you should use <div>
in place of a normal tag, such as <h1> or <p>.